+++ src/module/rules/rules.ts
@@ -7,6 +7,11 @@ import {PF2FixedProficiencyRuleElement} from "./elements/fixed-proficiency";
 import {PF2TempHPRuleElement} from "./elements/temphp";
 import {PF2DexterityModifierCapRuleElement} from "./elements/dexterity-modifier-cap";
 import {PF2DamageDiceRuleElement} from "./elements/damage-dice";
+import {PF2TogglePropertyRuleElement} from "./elements/toggle-property";
+import {PF2TokenImageRuleElement} from "./elements/token-image";
+import {PF2TokenSizeRuleElement} from "./elements/token-size";
+import {PF2BaseSpeedRuleElement} from "./elements/base-speed";
+import {PF2SenseRuleElement} from "./elements/sense";
 
 export {PF2RuleElement};
 
@@ -19,6 +24,11 @@ export class PF2RuleElements {
         'PF2E.RuleElement.FixedProficiency':  (ruleData, item) => new PF2FixedProficiencyRuleElement(ruleData, item),
         'PF2E.RuleElement.TempHP':  (ruleData, item) => new PF2TempHPRuleElement(ruleData, item),
         'PF2E.RuleElement.DamageDice': (ruleData, item) => new PF2DamageDiceRuleElement(ruleData, item),
+        'PF2E.RuleElement.ToggleProperty': (ruleData, item) => new PF2TogglePropertyRuleElement(ruleData, item),
+        'PF2E.RuleElement.TokenImage': (ruleData, item) => new PF2TokenImageRuleElement(ruleData, item),
+        'PF2E.RuleElement.TokenSize': (ruleData, item) => new PF2TokenSizeRuleElement(ruleData, item),
+        'PF2E.RuleElement.BaseSpeed': (ruleData, item) => new PF2BaseSpeedRuleElement(ruleData, item),
+        'PF2E.RuleElement.Sense': (ruleData, item) => new PF2SenseRuleElement(ruleData, item),
     });
 
     static custom: Record<string, (ruleData: PF2RuleElementData, item: ItemData) => PF2RuleElement> = {}